pyproject.toml
[project]
name = "wumlla"
version = "0.1.0"
description = "Discord Bot for local LLM"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.11.10",
    "discord-py>=2.4.0",
    "python-dotenv>=1.0.1",
]
